requirements.txt

fastapi[standard]
pymongo
motor
pydantic
python-dotenv
pyjwt
bcrypt
cloudinary